+++ packages/audioplayers/example/lib/player_widget.dart
@@ -115,7 +115,11 @@ class _PlayerWidgetState extends State<PlayerWidget> {
                 children: [
                   Slider(
                     onChanged: (v) {
-                      final Position = v * _duration!.inMilliseconds;
+                      final duration = _duration;
+                      if (duration == null) {
+                        return;
+                      }
+                      final Position = v * duration.inMilliseconds;
                       _audioPlayer
                           .seek(Duration(milliseconds: Position.round()));
                     },
@@ -225,11 +229,6 @@ class _PlayerWidgetState extends State<PlayerWidget> {
     if (result == 1) {
       setState(() => _playerState = PlayerState.PLAYING);
     }
-
-    // default playback rate is 1.0
-    // this should be called after _audioPlayer.play() or _audioPlayer.resume()
-    // this can also be called everytime the user wants to change playback rate in the UI
-    _audioPlayer.setPlaybackRate();
 
     return result;
   }